improve Auth component
